Fix default column type to decimal - better representation of php float

+++ src/Schema/Columns/Float_Column.php
@@ -60,7 +60,7 @@ class Float_Column extends Column implements Lengthable, Signable, Precisionable
 	 *
 	 * @var string
 	 */
-	protected string $type = Column_Types::FLOAT;
+	protected string $type = Column_Types::DECIMAL;
 
 	/**
 	 * The PHP type of the column.
